Fix error in micro-averaged evaluation metrics
No normalization was performed within each multi-class class variable.

+++ src/main/java/es/upm/fi/cig/multictbnc/performance/Metric.java
@@ -20,5 +20,5 @@ public interface Metric {
 	 * @param confusionMatrix a {@code Map} containing a confusion matrix
 	 * @return result of the evaluation metric.
 	 */
-	public double compute(Map<String, Integer> confusionMatrix);
+	public double compute(Map<String, Double> confusionMatrix);
 }
